Move some shared functionality upstream
