+++ noora/connectors/MssqlConnector.py
@@ -7,17 +7,6 @@ from noora.connectors.Connector import Connector
 
 
 class MssqlConnector(Connector):
-    def __init__(self):
-        Connector.__init__(self)
-        # FIXME: implement this upstream? MysqlConnector also uses it
-        self.__result = None
-
-    def get_result(self):
-        return self.__result
-
-    def set_result(self, result):
-        self.__result = result
-
     def execute(self, executable, properties):
         script = executable['script']
