+++ noora/connectors/MysqlConnector.py
@@ -7,17 +7,6 @@ from noora.processor.PreProcessor import PreProcessor
 
 
 class MysqlConnector(Connector):
-    def __init__(self):
-        Connector.__init__(self)
-        # FIXME: implement this upstream? MssqlConnector also uses it
-        self.__result = None
-
-    def get_result(self):
-        return self.__result
-
-    def set_result(self, result):
-        self.__result = result
-
     def execute(self, executable, properties):
         script = executable['script']
